Refactoring to Incorporate one column compositor.
